requirements-dev.txt
centerline
jupyter
geopandas
geopy
matplotlib
momepy
numba
opencv-python
osmnx
pillow
pyarrow
rasterio
scikit-image
scikit-learn
shapely
tqdm
torch
torchvision

argh
coolname
tabulate
tensorboardx
packaging
sphinx
torchaudio
runx
toolz
more-itertools
gdown
pytest
imageio